refactor: 💡 remove any type in export-file to csv API
