+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+48",
+  "version": "0.8.5+50",
   "private": false,
   "scripts": {
     "dev": "next dev",
